+++ src/runtime/testdata/testprog/syscalls.go
@@ -2,53 +2,10 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build linux
-
 package main
 
 import (
-	"bytes"
-	"fmt"
-	"io/ioutil"
-	"os"
-	"syscall"
+	"errors"
 )
 
-func gettid() int {
-	return syscall.Gettid()
-}
-
-func tidExists(tid int) (exists, supported bool) {
-	stat, err := ioutil.ReadFile(fmt.Sprintf("/proc/self/task/%d/stat", tid))
-	if os.IsNotExist(err) {
-		return false, true
-	}
-	// Check if it's a zombie thread.
-	state := bytes.Fields(stat)[2]
-	return !(len(state) == 1 && state[0] == 'Z'), true
-}
-
-func getcwd() (string, error) {
-	if !syscall.ImplementsGetwd {
-		return "", nil
-	}
-	// Use the syscall to get the current working directory.
-	// This is imperative for checking for OS thread state
-	// after an unshare since os.Getwd might just check the
-	// environment, or use some other mechanism.
-	var buf [4096]byte
-	n, err := syscall.Getcwd(buf[:])
-	if err != nil {
-		return "", err
-	}
-	// Subtract one for null terminator.
-	return string(buf[:n-1]), nil
-}
-
-func unshareFs() error {
-	return syscall.Unshare(syscall.CLONE_FS)
-}
-
-func chdir(path string) error {
-	return syscall.Chdir(path)
-}
+var errNotPermitted = errors.New("operation not permitted")
